Refactor: Factorize ListActivity hierarchy
